refactor: GuestMainPage -> GuestSpaceHomePage, SpaceHomePage -> HostSpaceHomePage로 변경
