+++ frontend/src/pages/host/spaceHomePage/HostSpaceHomePage.styles.ts
